+++ services/payment_service.go
@@ -1,6 +1,6 @@
 package services
 
 type PaymentServiceInterface interface {
-	ProcessPayment(fromEmail, toEmail string, amount float64) error
-	MerchantToBankPayment(merchantId, bankID string, amount float64) error
+	ProcessPayment(fromEmail, toEmail string, amount int64) error
+	MerchantToBankPayment(merchantId, bankID string, amount int64) error
 }
